+++ src/config/generalConfig.ts
@@ -13,13 +13,15 @@ verifyCriticalEnvVariable('JWT_REFRESH_SECRET');
 // Get the application environment code. This can differ from the NODE_ENV which bears special
 // meaning for Node applictions. For example when we deploy to the AWS development environment,
 // the NODE_ENV is `staging` but the APP_ENV is `dev`.
-const env = process.env.APP_ENV || 'dev';
+const env: string = process.env.APP_ENV || 'dev';
+
+const domain: string = process.env.DOMAIN;
 
 export const generalConfig = {
   restDataSourceCacheTtl: Number.parseInt(process.env.REST_DATA_SOURCE_CACHE_TTL) || 180,
 
   env,
-  domain: process.env.DOMAIN,
+  domain: domain,
   applicationName: env === 'prd' ? process.env.APP_NAME : `${process.env.APP_NAME} (${env})`,
   defaultAffiliatioURI: process.env.DEFAULT_AFFILIATION_URI,
   defaultSearchLimit: Number.parseInt(process.env.DEFAULT_SEARCH_LIMIT) || 20,