+++ src/types.ts
@@ -467,6 +467,14 @@ export type AffiliationLinkInput = {
   url: Scalars['String']['input'];
 };
 
+export type AffiliationLogoUpload = {
+  __typename?: 'AffiliationLogoUpload';
+  /** The fields that should be included in the body of the POST request to upload the logo (e.g. policy, signature, etc.) stored as a JSON string */
+  fields: Scalars['String']['output'];
+  /** The URL to which the affiliation logo should be uploaded */
+  url: Scalars['String']['output'];
+};
+
 /** The provenance of an Affiliation record */
 export type AffiliationProvenance =
   /** Created and managed within the DMPTool */
@@ -1460,8 +1468,10 @@ export type Mutation = {
   createTemplateVersion?: Maybe<Template>;
   /** Deactivate the specified user Account (Admin only) */
   deactivateUser?: Maybe<User>;
-  /** Generate a presigned URL to upload an affiliation logo to the CloudFront CDN S3 bucket */
-  generateLogoUploadURL?: Maybe<Scalars['String']['output']>;
+  /** Finalizes the upload of an affiliation logo to the CloudFront CDN S3 bucket. The logoName should equal the 'key' from the fields returned by the generateLogoUploadURL mutation. */
+  finalizeLogoUpload: Affiliation;
+  /** Generate a presigned URL to upload an affiliation logo to the CloudFront CDN S3 bucket. The URL and fields returned are used to upload the logo to S3. */
+  generateLogoUploadURL?: Maybe<AffiliationLogoUpload>;
   /** The custom guidance for the custom section */
   guidance?: Maybe<Scalars['String']['output']>;
   /** The introduction to the custom section */
@@ -1566,8 +1576,6 @@ export type Mutation = {
   unpublishTemplateCustomization: TemplateCustomizationOverview;
   /** Update an Affiliation */
   updateAffiliation?: Maybe<Affiliation>;
-  /** Update the logo URL for an affiliation */
-  updateAffiliationLogo?: Maybe<Affiliation>;
   /** Edit an answer */
   updateAnswer?: Maybe<Answer>;
   /** Update comment for an answer  */
@@ -1879,8 +1887,15 @@ export type MutationDeactivateUserArgs = {
 };
 
 
+export type MutationFinalizeLogoUploadArgs = {
+  affiliationURI: Scalars['String']['input'];
+  logoName: Scalars['String']['input'];
+};
+
+
 export type MutationGenerateLogoUploadUrlArgs = {
   affiliationURI: Scalars['String']['input'];
+  contentType: Scalars['String']['input'];
   fileName: Scalars['String']['input'];
 };
 
@@ -2128,12 +2143,6 @@ export type MutationUnpublishTemplateCustomizationArgs = {
 
 export type MutationUpdateAffiliationArgs = {
   input: AffiliationInput;
-};
-
-
-export type MutationUpdateAffiliationLogoArgs = {
-  accessURL: Scalars['String']['input'];
-  affiliationURI: Scalars['String']['input'];
 };
 
 
@@ -5943,6 +5952,7 @@ export type ResolversTypes = {
   AffiliationInput: AffiliationInput;
   AffiliationLink: ResolverTypeWrapper<AffiliationLink>;
   AffiliationLinkInput: AffiliationLinkInput;
+  AffiliationLogoUpload: ResolverTypeWrapper<AffiliationLogoUpload>;
   AffiliationProvenance: AffiliationProvenance;
   AffiliationSearch: ResolverTypeWrapper<AffiliationSearch>;
   AffiliationSearchResults: ResolverTypeWrapper<AffiliationSearchResults>;
@@ -6192,6 +6202,7 @@ export type ResolversParentTypes = {
   AffiliationInput: AffiliationInput;
   AffiliationLink: AffiliationLink;
   AffiliationLinkInput: AffiliationLinkInput;
+  AffiliationLogoUpload: AffiliationLogoUpload;
   AffiliationSearch: AffiliationSearch;
   AffiliationSearchResults: AffiliationSearchResults;
   AlternateIdentifier: AlternateIdentifier;
@@ -6457,6 +6468,11 @@ export type AffiliationErrorsResolvers<ContextType = MyContext, ParentType exten
 export type AffiliationLinkResolvers<ContextType = MyContext, ParentType extends ResolversParentTypes['AffiliationLink'] = ResolversParentTypes['AffiliationLink']> = {
   id?: Resolver<ResolversTypes['Int'], ParentType, ContextType>;
   text?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
+  url?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
+};
+
+export type AffiliationLogoUploadResolvers<ContextType = MyContext, ParentType extends ResolversParentTypes['AffiliationLogoUpload'] = ResolversParentTypes['AffiliationLogoUpload']> = {
+  fields?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
   url?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
 };
 
@@ -6965,7 +6981,8 @@ export type MutationResolvers<ContextType = MyContext, ParentType extends Resolv
   completeFeedback?: Resolver<Maybe<ResolversTypes['PlanFeedback']>, ParentType, ContextType, RequireFields<MutationCompleteFeedbackArgs, 'planFeedbackId' | 'planId'>>;
   createTemplateVersion?: Resolver<Maybe<ResolversTypes['Template']>, ParentType, ContextType, RequireFields<MutationCreateTemplateVersionArgs, 'latestPublishVisibility' | 'templateId'>>;
   deactivateUser?: Resolver<Maybe<ResolversTypes['User']>, ParentType, ContextType, RequireFields<MutationDeactivateUserArgs, 'userId'>>;
-  generateLogoUploadURL?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType, RequireFields<MutationGenerateLogoUploadUrlArgs, 'affiliationURI' | 'fileName'>>;
+  finalizeLogoUpload?: Resolver<ResolversTypes['Affiliation'], ParentType, ContextType, RequireFields<MutationFinalizeLogoUploadArgs, 'affiliationURI' | 'logoName'>>;
+  generateLogoUploadURL?: Resolver<Maybe<ResolversTypes['AffiliationLogoUpload']>, ParentType, ContextType, RequireFields<MutationGenerateLogoUploadUrlArgs, 'affiliationURI' | 'contentType' | 'fileName'>>;
   guidance?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
   introduction?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
   markAsDefaultTemplate?: Resolver<Maybe<ResolversTypes['Template']>, ParentType, ContextType, RequireFields<MutationMarkAsDefaultTemplateArgs, 'templateId'>>;
@@ -7018,7 +7035,6 @@ export type MutationResolvers<ContextType = MyContext, ParentType extends Resolv
   unpublishGuidanceGroup?: Resolver<ResolversTypes['GuidanceGroup'], ParentType, ContextType, RequireFields<MutationUnpublishGuidanceGroupArgs, 'guidanceGroupId'>>;
   unpublishTemplateCustomization?: Resolver<ResolversTypes['TemplateCustomizationOverview'], ParentType, ContextType, RequireFields<MutationUnpublishTemplateCustomizationArgs, 'templateCustomizationId'>>;
   updateAffiliation?: Resolver<Maybe<ResolversTypes['Affiliation']>, ParentType, ContextType, RequireFields<MutationUpdateAffiliationArgs, 'input'>>;
-  updateAffiliationLogo?: Resolver<Maybe<ResolversTypes['Affiliation']>, ParentType, ContextType, RequireFields<MutationUpdateAffiliationLogoArgs, 'accessURL' | 'affiliationURI'>>;
   updateAnswer?: Resolver<Maybe<ResolversTypes['Answer']>, ParentType, ContextType, RequireFields<MutationUpdateAnswerArgs, 'answerId'>>;
   updateAnswerComment?: Resolver<Maybe<ResolversTypes['AnswerComment']>, ParentType, ContextType, RequireFields<MutationUpdateAnswerCommentArgs, 'answerCommentId' | 'answerId' | 'commentText'>>;
   updateCustomQuestion?: Resolver<ResolversTypes['CustomQuestion'], ParentType, ContextType, RequireFields<MutationUpdateCustomQuestionArgs, 'input'>>;
@@ -8474,6 +8490,7 @@ export type Resolvers<ContextType = MyContext> = {
   AffiliationEmailDomain?: AffiliationEmailDomainResolvers<ContextType>;
   AffiliationErrors?: AffiliationErrorsResolvers<ContextType>;
   AffiliationLink?: AffiliationLinkResolvers<ContextType>;
+  AffiliationLogoUpload?: AffiliationLogoUploadResolvers<ContextType>;
   AffiliationSearch?: AffiliationSearchResolvers<ContextType>;
   AffiliationSearchResults?: AffiliationSearchResultsResolvers<ContextType>;
   AlternateIdentifier?: AlternateIdentifierResolvers<ContextType>;